refactor: change mocked client file name to better understand what it exports
